Extract todo specific logic to todo blueprint
